backend-app/requirements.txt
Flask
mysql-connector-python
Flask-Cors # Tambahkan ini untuk menangani CORS
